requirements.txt
beautifulsoup4==4.11.2
Flask==2.0.2
mysql_connector_repackaged==0.3.1
n2yo==0.0.5
n2yo_api==0.1.3
pandas==1.1.5
requests==2.26.0